requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
pyarrow>=17.0.0
duckdb>=1.1.3
snowflake-connector-python>=3.12.2
